refactor(multi_object_tracker): consolidate object model types into a single header
- Replaced multiple includes of object_model/types.hpp with a unified include of types.hpp across various files to streamline dependencies.
- Introduced a new types.hpp file that consolidates type definitions previously scattered across different files, improving organization and maintainability.
- Updated CMakeLists.txt to reflect the new file structure.
