+++ perception/autoware_multi_object_tracker/include/autoware/multi_object_tracker/configurations.hpp
@@ -15,7 +15,7 @@
 #ifndef AUTOWARE__MULTI_OBJECT_TRACKER__CONFIGURATIONS_HPP_
 #define AUTOWARE__MULTI_OBJECT_TRACKER__CONFIGURATIONS_HPP_
 
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 
 #include <cstddef>
 #include <functional>
