+++ perception/autoware_multi_object_tracker/include/autoware/multi_object_tracker/tracker/model/polygon_tracker.hpp
@@ -16,7 +16,7 @@
 #define AUTOWARE__MULTI_OBJECT_TRACKER__TRACKER__MODEL__POLYGON_TRACKER_HPP_
 
 #include "autoware/multi_object_tracker/object_model/object_model.hpp"
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/tracker/model/tracker_base.hpp"
 #include "autoware/multi_object_tracker/tracker/motion_model/cv_motion_model.hpp"
 #include "autoware/multi_object_tracker/tracker/motion_model/static_motion_model.hpp"
