+++ perception/autoware_multi_object_tracker/include/autoware/multi_object_tracker/tracker/shape_model/unstable_shape_filter.hpp
@@ -15,7 +15,7 @@
 #ifndef AUTOWARE__MULTI_OBJECT_TRACKER__TRACKER__SHAPE_MODEL__UNSTABLE_SHAPE_FILTER_HPP_
 #define AUTOWARE__MULTI_OBJECT_TRACKER__TRACKER__SHAPE_MODEL__UNSTABLE_SHAPE_FILTER_HPP_
 
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 
 #include <Eigen/Core>
