+++ perception/autoware_multi_object_tracker/lib/tracker/model/tracker_base.cpp
@@ -14,7 +14,7 @@
 
 #include "autoware/multi_object_tracker/tracker/model/tracker_base.hpp"
 
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/object_model/uuid.hpp"
 
 #include <autoware_utils_geometry/geometry.hpp>
