+++ perception/autoware_multi_object_tracker/src/debugger/debug_object.hpp
@@ -15,7 +15,7 @@
 #ifndef DEBUGGER__DEBUG_OBJECT_HPP_
 #define DEBUGGER__DEBUG_OBJECT_HPP_
 
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/tracker/model/tracker_base.hpp"
 
 #include <rclcpp/rclcpp.hpp>
