+++ perception/autoware_multi_object_tracker/src/multi_object_tracker_core.hpp
@@ -16,7 +16,7 @@
 #define MULTI_OBJECT_TRACKER_CORE_HPP_
 
 #include "autoware/multi_object_tracker/association/bev_association.hpp"
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/odometry.hpp"
 #include "debugger/debugger.hpp"
 #include "processor/input_manager.hpp"
