+++ perception/autoware_multi_object_tracker/src/processor/input_manager.hpp
@@ -15,7 +15,7 @@
 #ifndef PROCESSOR__INPUT_MANAGER_HPP_
 #define PROCESSOR__INPUT_MANAGER_HPP_
 
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/odometry.hpp"
 #include "rclcpp/rclcpp.hpp"
