+++ perception/autoware_multi_object_tracker/src/processor/processor.cpp
@@ -16,7 +16,7 @@
 
 #include "autoware/multi_object_tracker/object_model/object_model.hpp"
 #include "autoware/multi_object_tracker/object_model/shapes.hpp"
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/tracker/tracker.hpp"
 
 #include <autoware_perception_msgs/msg/tracked_objects.hpp>
