+++ perception/autoware_multi_object_tracker/src/processor/processor.hpp
@@ -18,7 +18,7 @@
 #include "autoware/multi_object_tracker/association/adaptive_threshold_cache.hpp"
 #include "autoware/multi_object_tracker/association/association_manager.hpp"
 #include "autoware/multi_object_tracker/association/tracker_overlap_manager.hpp"
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/tracker/model/tracker_base.hpp"
 
 #include <autoware_utils_debug/time_keeper.hpp>
