+++ perception/autoware_multi_object_tracker/test/test_bench.hpp
@@ -15,7 +15,7 @@
 #define TEST_BENCH_HPP_
 #include "../src/multi_object_tracker_node.hpp"
 #include "autoware/multi_object_tracker/object_model/shapes.hpp"
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/tracker/model/tracker_base.hpp"
 #include "autoware/multi_object_tracker/uncertainty/uncertainty_processor.hpp"
 #include "test_utils.hpp"
