+++ perception/autoware_multi_object_tracker/test/test_multi_object_tracker.cpp
@@ -13,7 +13,7 @@
 // limitations under the License.
 #include "../src/multi_object_tracker_node.hpp"
 #include "autoware/multi_object_tracker/object_model/shapes.hpp"
-#include "autoware/multi_object_tracker/object_model/types.hpp"
+#include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/odometry.hpp"
 #include "autoware/multi_object_tracker/uncertainty/uncertainty_processor.hpp"
 #include "test_bench.hpp"
