+++ perception/autoware_multi_object_tracker/test/test_utils.hpp
@@ -16,7 +16,7 @@
 #ifndef TEST_UTILS_HPP_
 #define TEST_UTILS_HPP_
 
-#include <autoware/multi_object_tracker/object_model/types.hpp>
+#include <autoware/multi_object_tracker/types.hpp>
 #include <rosbag2_cpp/reader.hpp>
 #include <rosbag2_cpp/writer.hpp>
 #include <rosbag2_cpp/writers/sequential_writer.hpp>
